Refactored TokenFragment
